+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Model/CostInfomation.swift
@@ -1,0 +1,27 @@
+//
+//  CostInfomation.swift
+//  ParcelInvoiceMaker
+//
+//  Created by Kyeongmo Yang on 3/7/24.
+//
+
+import Foundation
+
+enum Discount: Int {
+    case none = 0, vip, coupon
+}
+
+struct CostInfomation {
+    let deliveryCost: Int
+    private let discount: Discount
+    var discountedCost: Int {
+        switch discount {
+        case .none:
+            return deliveryCost
+        case .vip:
+            return deliveryCost / 5 * 4
+        case .coupon:
+            return deliveryCost / 2
+        }
+    }
+}
