+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Model/DatabaseParcelInformationPersistence.swift
@@ -1,0 +1,15 @@
+//
+//  DatabaseParcelInformationPersistence.swift
+//  ParcelInvoiceMaker
+//
+//  Created by Kyeongmo Yang on 3/7/24.
+//
+
+import Foundation
+
+class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
+    func save(parcelInformation: ParcelInformation) {
+        // 데이터베이스에 주문 정보 저장
+        print("발송 정보를 데이터베이스에 저장했습니다.")
+    }
+}
