+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Model/ParcelInfomation.swift
@@ -18,27 +18,3 @@ class ParcelInformation {
         self.costInfomation = costInfomation
     }
 }
-
-struct ReceiverInfomation {
-    var receiverName: String
-    var receiverMobile: String
-}
-
-struct CostInfomation {
-    let deliveryCost: Int
-    private let discount: Discount
-    var discountedCost: Int {
-        switch discount {
-        case .none:
-            return deliveryCost
-        case .vip:
-            return deliveryCost / 5 * 4
-        case .coupon:
-            return deliveryCost / 2
-        }
-    }
-}
-
-enum Discount: Int {
-    case none = 0, vip, coupon
-}
