+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Model/ParcelProcessor.swift
@@ -1,39 +1,12 @@
 //
-//  ParcelInvoiceMaker - ParcelProcessor.swift
-//  Created by yagom.
-//  Copyright © yagom. All rights reserved.
+//  ParcelProcessor.swift
+//  ParcelInvoiceMaker
+//
+//  Created by Kyeongmo Yang on 3/7/24.
 //
 
 import Foundation
 
-protocol ParcelProcessorProtocol {
+protocol ParcelProcessor {
     func process(parcelInformation: ParcelInformation, onComplete: (ParcelInformation) -> Void)
-}
-
-class ParcelOrderProcessor: ParcelProcessorProtocol {
-    let databaseParcelInformationPersistence: ParcelInformationPersistence
-    
-    init(databaseParcelInformationPersistence: ParcelInformationPersistence) {
-        self.databaseParcelInformationPersistence = databaseParcelInformationPersistence
-    }
-    
-    // 택배 주문 처리 로직
-    func process(parcelInformation: ParcelInformation, onComplete: (ParcelInformation) -> Void) {
-        
-        // 데이터베이스에 주문 저장
-        databaseParcelInformationPersistence.save(parcelInformation: parcelInformation)
-        
-        onComplete(parcelInformation)
-    }
-}
-
-class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation) {
-        // 데이터베이스에 주문 정보 저장
-        print("발송 정보를 데이터베이스에 저장했습니다.")
-    }
-}
-
-protocol ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation)
 }
