+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Model/ReceiverInfomation.swift
@@ -1,0 +1,13 @@
+//
+//  ReceiverInfomation.swift
+//  ParcelInvoiceMaker
+//
+//  Created by Kyeongmo Yang on 3/7/24.
+//
+
+import Foundation
+
+struct ReceiverInfomation {
+    var name: String
+    var mobile: String
+}
